Added interfaces for future LxAllocator class
