+++ torch_spyre/_inductor/layout_backend.py
@@ -1,0 +1,36 @@
+# Copyright 2025 The Torch-Spyre Authors.
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
+
+
+from dataclasses import dataclass
+from typing import list, Optional, dict, str, float
+from abc import abstractmethod
+
+
+@dataclass
+class LifetimeBoundBuffer:
+    name: str
+    size: int
+    start_time: int
+    end_time: int
+    address: Optional[int] = None
+    spilled: Optional[bool] = None
+    heuristic: dict[str, float] = {}
+
+
+class LayoutSolver:
+    @abstractmethod
+    def plan_layout(self, buffers: list[LifetimeBoundBuffer]) -> list[LifetimeBoundBuffer]:
+        pass
+
